Stop submitting metrics across Instance Name and ID dimensions
Prior to this change, several metrics were submitted with both the
InstanceId and InstanceName dimensions. This prevents us from analysing
metrics only on the InstanceName dimension (which is the more interesting one).

This change stop including the InstanceId dimension as it's not that
interesting and allow the metrics for dashboards to work correctly.