Force a dist-upgrade instead of apt-get upgrade
The dist-upgrade is a superset of upgrade. Apart from the upgrade function,
it can also intelligently handle changes in the package dependencies.
This includes removing dependency packages that are no longer necessary or
resolving conflicts between packages that arose because of changes in the dependencies.

This should solve most kernel CVEs flagged by Inspector